+++ src/core/AxiosErrorCodes.ts
@@ -1,0 +1,5 @@
+enum AxiosErrorCodes {
+    ERR_NETWORK = "ERR_NETWORK"
+}
+
+export default AxiosErrorCodes;